Move wait for tun up event into `TalpidVpnService`
